+++ custom_deps/run.sh
@@ -1,33 +1,37 @@
-#!/bin/bash
+#!/usr/bin/env bashio
 set -e
 
-CONFIG_PATH=/data/options.json
-
-PYPI="$(jq --raw-output '.pypi | join(" ")' ${CONFIG_PATH})"
-APK="$(jq --raw-output '.apk | join(" ") // empty' ${CONFIG_PATH})"
+PYPI=$(bashio::config 'pypi')
+APK=$(bashio::config 'apk')
 
 # Cleanup old deps
-echo "[Info] Remove old deps"
+bashio::log.info "Remove old deps"
 rm -rf /config/deps/*
 
 # Need custom apk for build?
 if [ -n "${APK}" ]; then
-    echo "[Info] Install apks for build"
-    # shellcheck disable=SC2086
-    if ! ERROR="$(apk add --no-cache ${APK})"; then
-        echo "[Error] Can't install packages!"
-        echo "${ERROR}" && exit 1
-    fi
+    for dep in ${APK}; do
+        bashio::log.info "Installing alpine package '${dep}'"
+        # shellcheck disable=SC2086
+        if ! ERROR="$(apk add --no-cache ${dep})"; then
+            bashio::log.error "Can't install ${dep}!"
+            bashio::log.error "${ERROR}" && exit 1
+        fi
+    done
 fi
 
 # Install pypi modules
-echo "[Info] Install pypi modules into deps"
+bashio::log.info "Install pypi modules into deps"
 export PYTHONUSERBASE=/config/deps
 
-# shellcheck disable=SC2086
-if ! ERROR="$(pip3 install --user --no-cache-dir --prefix= --no-dependencies ${PYPI})"; then
-    echo "[Error] Can't install PyPI packages!"
-    echo "${ERROR}" && exit 1
-fi
+for dep in ${PYPI}; do
+    bashio::log.info "Installing python package '${dep}'"
+    
+    # shellcheck disable=SC2086
+    if ! ERROR="$(pip3 install --user --no-cache-dir --prefix= --no-dependencies --disable-pip-version-check ${dep})"; then
+        bashio::log.error "Can't install ${dep}!"
+        bashio::log.error "${ERROR}" && exit 1
+    fi
+done
 
-echo "[Info] done"
+bashio::log.info "Done"