+++ custom_deps/CHANGELOG.md
@@ -1,13 +1,22 @@
 # Changelog
 
-# 1.3.0
+## 1.3.1
+
+- Add `--disable-pip-version-check` to the pip install command
+- Use bashio to get configuration
+- Use bashio for logs
+- Log each package that is being installed
+
+## 1.3.0
+
 - Update Python to 3.9
 - Set minimum Core version to 2021.7.0
-# 1.2.0
+
+## 1.2.0
 
 - Fix build
 
-# 1.1.0
+## 1.1.0
 
 - Update Python 3.8